pyproject.toml
[project]
name = "noravue"
version = "0.1.0"
description = "Screenshot analysis and triage system"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "spacy>=3.8.5",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3"
]

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
